+++ CHANGELOG.md
@@ -2,6 +2,7 @@ v7 ???:
 * Load the "My Shows" list incrementally, so even long lists of shows appear
   quickly.  More are loaded only when they scroll into view.
 * Do not try to schedule a Season Pass with no channel.
+* Discover, save, and re-use the "bodyId" value that will soon be required.
 
 v6 Aug 23, 2011:
 * Fix a stupid mistake I made in v5.